+++ app/Http/Controllers/TagController.php
@@ -6,11 +6,17 @@ use App\Http\Exeptions\BadRequestException;
 use App\Http\Exeptions\NotFoundException;
 use App\Http\Resources\TagResource;
 use App\Models\Tag;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
 class TagController extends Controller
 {
+  public function __construct()
+  {
+    $this->middleware('auth.jwt')->except('index');
+  }
+
   public function index(): JsonResponse
   {
     try {
@@ -35,6 +41,7 @@ class TagController extends Controller
   public function store(Request $request): JsonResponse
   {
     try {
+      $this->authorize('create', Tag::class);
       $validator = validator($request->all(), [
         'name' => 'required|string|max:255'
       ]);
@@ -44,6 +51,8 @@ class TagController extends Controller
       $tag = Tag::create([
         'name' => $request->input('name')
       ]);
+    } catch (AuthorizationException $e) {
+      return response()->json(['error' => $e->getMessage()], 403);
     } catch (BadRequestException $e) {
       return response()->json($e->getError(), $e->getCode());
     } catch (\Throwable $e) {
@@ -55,6 +64,7 @@ class TagController extends Controller
   public function update(Request $request, $id): JsonResponse
   {
     try {
+      $this->authorize('update', Tag::class);
       $validator = validator($request->all(), [
         'name' => 'nullable|string|max:255'
       ]);
@@ -67,12 +77,12 @@ class TagController extends Controller
       $tag->update([
         'name' => $request->input('name') ?? $tag->name
       ]);
+    } catch (AuthorizationException $e) {
+      return response()->json(['error' => $e->getMessage()], 403);
     } catch (BadRequestException $e) {
       return response()->json($e->getError(), $e->getCode());
     } catch (NotFoundException $e) {
       return response()->json($e->getError(), $e->getCode());
-    } catch (\Throwable $e) {
-      return response()->json(['error' => $e->getMessage()], 500);
     }
     return response()->json(new TagResource($tag));
   }
@@ -80,10 +90,13 @@ class TagController extends Controller
   public function destroy($id): JsonResponse
   {
     try {
+      $this->authorize('delete', Tag::class);
       if (!$tag = Tag::find($id)) {
         throw new NotFoundException('Tag not found');
       }
       $tag->delete();
+    } catch (AuthorizationException $e) {
+      return response()->json(['error' => $e->getMessage()], 403);
     } catch (NotFoundException $e) {
       return response()->json($e->getError(), $e->getCode());
     } catch (\Throwable $e) {
